Add OpenMP support for Fortran builds
All saved variables in the Fortran 95 routines are now declared as

!$OMP THREADPRIVATE(...)

make fortran-mp, fortran2-mp, and fortran3-mp create an archive with
the name libSHTOOLS-mp.a. This can be used in a thread safe manner
within an OpenMP parellized construct. make fortran-tests-mp compiles
and runs the fortan routines when linking to the SHTOOLS-mp library.