Fix CRITICAL (BUT RARE) bug with MakeGridDH, MakeGravGridDH, MakeMagGridDH and MakeGravGradGridDH when sampling = 1
In the above routines, the rows of the output grid are calculated by inverse
Fourier transmforming a vector that depends upon the spherical harmonic
coefficients. This vector, when using the complex-to-real FFTW routines,
includes one element that corresponds to the coefficients with m=lmax+1 (for reasons
that I don't fully understand...). This element of the array was not properly
initialized to zero in the Fortran code, and if the Fortran compiler did not
explictly zero all new arrays, this could have resulted in incorrect output.
Given that this term is 1 element above the Nyquist frequency, corresponding to
m=lmax, if this element were not zero, each column of the grid would contain a
component that oscillates from -1 to +1 (sclaed by the magnitude of the element).

When the element in the array is not properly initiallized, the routine gives
wildly inaccurate results, as the element often has an extremely large value.
Nevertheless, since this corresponds to a frequency that is beyond the Nyquist frequency, a
spherical harmonic expansion of this grid will often give correct values up to
lmax. Routines that are based on GLQ grids, as well as DH grids with
sampling = 2, are not affected.