Fix wrapping of Fortran subroutines
Currently, docs for wrapped subroutines start with:
```
Help on function wrapped_func in module pyshtools.backends.shtools:

wrapped_func(*args, **kwargs)
```

This is because only `__doc__` is copied, but not the `__name__` and
other special properties. Using `functools.wraps` copies all important
properties, and changes the docs to:
```
Help on function BAtoHilmDH in module pyshtools.backends.shtools:

BAtoHilmDH(...)
```

But note there needs to be a small tweak to `__name__` because f2py adds
an extra 'function ' prefix to the name.